==> Text Project -In God we trust (P7)/Notebooks/1 - Scraping.ipynb <==
# import libraries
import snscrape.modules.twitter as sntwitter
import pandas as pd

# scrape from specific users in a specific timeline, english language and no retweet
accounts_df = pd.read_csv("path/to/(non)religion.csv", sep = ';')
accounts = accounts_df['Handle'].tolist()

authors = []
attributes_container = []

str1 = 'from:'
str2 = ' since:2022-03-01 until:2022-05-31 -is:retweet lang:en'
for account in accounts:
    for i,tweet in enumerate(sntwitter.TwitterSearchScraper(str1 + str(account) +str2).get_items()):
        if i>250:
            attributes_container.append([tweet.date, tweet.likeCount, tweet.sourceLabel, tweet.content])
            authors.append(account)
            continue
        

tweets_df = pd.DataFrame(attributes_container, columns=["Date Created", "Number of Likes", "Source of Tweet", "Tweet"])

tweets_df['Authors'] = authors

#save new version of df
tweets_df.to_csv(r"path/to/Tweet_update.csv", 
                index = False)
